+++ tests/ssh/test_parallel_client.py
@@ -232,10 +232,8 @@ class LibSSHParallelTest(unittest.TestCase):
                                    pkey=self.user_key,
                                    timeout=client_timeout,
                                    num_retries=1)
-        cmd = spawn(client.run_command, 'sleep 1', stop_on_errors=False)
-        output = cmd.get(timeout=client_timeout * 1000)
-        self.assertIsInstance(output[0].exception,
-                              ConnectionErrorException)
+        output = client.run_command('sleep 1', stop_on_errors=False)
+        self.assertIsInstance(output[0].exception, ConnectionErrorException)
 
     def test_zero_timeout(self):
         host = '127.0.0.2'